DTD-4457 combined the then steps in SA and suppression
