+++ src/test/scala/uk/gov/hmrc/test/api/scalatest/steps/helpers/sol/StatementOfLiabilityStepHelpers.scala
@@ -56,83 +56,83 @@ trait StatementOfLiabilityStepHelpers { this: Matchers =>
     context.headers = response.headers.map { case (key, values) => key -> values.headOption.getOrElse("") }
   }
 
-  // ^service returns debt statement of liability data$
-  def serviceReturnsDebtStatementOfLiabilityData(
-    context: StatementOfLiabilityContext,
-    expectedResponse: SolCalculationSummaryResponse
-  ): Unit = {
-    val actual = context.responseBody
-    println(s"actualResponseBody : " + actual)
-    println(s"expectedResponse : " + Some(expectedResponse))
-
-    context.status shouldBe 200
-
-    actual match {
-      case Some(actual) =>
-        withClue("amountIntTotal: ") {
-          actual.amountIntTotal shouldBe expectedResponse.amountIntTotal
-        }
-        withClue("combinedDailyAccrual: ") {
-          actual.combinedDailyAccrual shouldBe expectedResponse.combinedDailyAccrual
-        }
-        withClue("debts list length: ") {
-          actual.debts.length shouldBe expectedResponse.debts.length
-        }
-
-        // Verify each SolCalculation in debts list
-        actual.debts.zip(expectedResponse.debts).zipWithIndex.foreach { case ((actualDebt, expectedDebt), debtIndex) =>
-          withClue(s"debts[$debtIndex].debtId: ") {
-            actualDebt.debtId shouldBe expectedDebt.debtId
-          }
-          withClue(s"debts[$debtIndex].mainTrans: ") {
-            actualDebt.mainTrans shouldBe expectedDebt.mainTrans
-          }
-          withClue(s"debts[$debtIndex].debtTypeDescription: ") {
-            actualDebt.debtTypeDescription shouldBe expectedDebt.debtTypeDescription
-          }
-          withClue(s"debts[$debtIndex].interestDueDebtTotal: ") {
-            actualDebt.interestDueDebtTotal shouldBe expectedDebt.interestDueDebtTotal
-          }
-          withClue(s"debts[$debtIndex].totalAmountIntDebt: ") {
-            actualDebt.totalAmountIntDebt shouldBe expectedDebt.totalAmountIntDebt
-          }
-          withClue(s"debts[$debtIndex].combinedDailyAccrual: ") {
-            actualDebt.combinedDailyAccrual shouldBe expectedDebt.combinedDailyAccrual
-          }
-          withClue(s"debts[$debtIndex].parentMainTrans: ") {
-            actualDebt.parentMainTrans shouldBe expectedDebt.parentMainTrans
-          }
-          withClue(s"debts[$debtIndex].duties list length: ") {
-            actualDebt.duties.length shouldBe expectedDebt.duties.length
-          }
-
-          // Verify each SolDuty in duties list
-          actualDebt.duties.zip(expectedDebt.duties).zipWithIndex.foreach {
-            case ((actualDuty, expectedDuty), dutyIndex) =>
-              withClue(s"debts[$debtIndex].duties[$dutyIndex].subTrans: ") {
-                actualDuty.subTrans shouldBe expectedDuty.subTrans
-              }
-              withClue(s"debts[$debtIndex].duties[$dutyIndex].dutyTypeDescription: ") {
-                actualDuty.dutyTypeDescription shouldBe expectedDuty.dutyTypeDescription
-              }
-              withClue(s"debts[$debtIndex].duties[$dutyIndex].unpaidAmountDuty: ") {
-                actualDuty.unpaidAmountDuty shouldBe expectedDuty.unpaidAmountDuty
-              }
-              withClue(s"debts[$debtIndex].duties[$dutyIndex].combinedDailyAccrual: ") {
-                actualDuty.combinedDailyAccrual shouldBe expectedDuty.combinedDailyAccrual
-              }
-              withClue(s"debts[$debtIndex].duties[$dutyIndex].interestBearing: ") {
-                actualDuty.interestBearing shouldBe expectedDuty.interestBearing
-              }
-              withClue(s"debts[$debtIndex].duties[$dutyIndex].interestOnlyIndicator: ") {
-                actualDuty.interestOnlyIndicator shouldBe expectedDuty.interestOnlyIndicator
-              }
-          }
-        }
-      case None         => fail("Response body is empty")
-    }
-  }
-
+//  // ^service returns debt statement of liability data$
+//  def serviceReturnsDebtStatementOfLiabilityData(
+//    context: StatementOfLiabilityContext,
+//    expectedResponse: SolCalculationSummaryResponse
+//  ): Unit = {
+//    val actual = context.responseBody
+//    println(s"actualResponseBody : " + actual)
+//    println(s"expectedResponse : " + Some(expectedResponse))
+//
+//    context.status shouldBe 200
+//
+//    actual match {
+//      case Some(actual) =>
+//        withClue("amountIntTotal: ") {
+//          actual.amountIntTotal shouldBe expectedResponse.amountIntTotal
+//        }
+//        withClue("combinedDailyAccrual: ") {
+//          actual.combinedDailyAccrual shouldBe expectedResponse.combinedDailyAccrual
+//        }
+//        withClue("debts list length: ") {
+//          actual.debts.length shouldBe expectedResponse.debts.length
+//        }
+//
+//        // Verify each SolCalculation in debts list
+//        actual.debts.zip(expectedResponse.debts).zipWithIndex.foreach { case ((actualDebt, expectedDebt), debtIndex) =>
+//          withClue(s"debts[$debtIndex].debtId: ") {
+//            actualDebt.debtId shouldBe expectedDebt.debtId
+//          }
+//          withClue(s"debts[$debtIndex].mainTrans: ") {
+//            actualDebt.mainTrans shouldBe expectedDebt.mainTrans
+//          }
+//          withClue(s"debts[$debtIndex].debtTypeDescription: ") {
+//            actualDebt.debtTypeDescription shouldBe expectedDebt.debtTypeDescription
+//          }
+//          withClue(s"debts[$debtIndex].interestDueDebtTotal: ") {
+//            actualDebt.interestDueDebtTotal shouldBe expectedDebt.interestDueDebtTotal
+//          }
+//          withClue(s"debts[$debtIndex].totalAmountIntDebt: ") {
+//            actualDebt.totalAmountIntDebt shouldBe expectedDebt.totalAmountIntDebt
+//          }
+//          withClue(s"debts[$debtIndex].combinedDailyAccrual: ") {
+//            actualDebt.combinedDailyAccrual shouldBe expectedDebt.combinedDailyAccrual
+//          }
+//          withClue(s"debts[$debtIndex].parentMainTrans: ") {
+//            actualDebt.parentMainTrans shouldBe expectedDebt.parentMainTrans
+//          }
+//          withClue(s"debts[$debtIndex].duties list length: ") {
+//            actualDebt.duties.length shouldBe expectedDebt.duties.length
+//          }
+//
+//          // Verify each SolDuty in duties list
+//          actualDebt.duties.zip(expectedDebt.duties).zipWithIndex.foreach {
+//            case ((actualDuty, expectedDuty), dutyIndex) =>
+//              withClue(s"debts[$debtIndex].duties[$dutyIndex].subTrans: ") {
+//                actualDuty.subTrans shouldBe expectedDuty.subTrans
+//              }
+//              withClue(s"debts[$debtIndex].duties[$dutyIndex].dutyTypeDescription: ") {
+//                actualDuty.dutyTypeDescription shouldBe expectedDuty.dutyTypeDescription
+//              }
+//              withClue(s"debts[$debtIndex].duties[$dutyIndex].unpaidAmountDuty: ") {
+//                actualDuty.unpaidAmountDuty shouldBe expectedDuty.unpaidAmountDuty
+//              }
+//              withClue(s"debts[$debtIndex].duties[$dutyIndex].combinedDailyAccrual: ") {
+//                actualDuty.combinedDailyAccrual shouldBe expectedDuty.combinedDailyAccrual
+//              }
+//              withClue(s"debts[$debtIndex].duties[$dutyIndex].interestBearing: ") {
+//                actualDuty.interestBearing shouldBe expectedDuty.interestBearing
+//              }
+//              withClue(s"debts[$debtIndex].duties[$dutyIndex].interestOnlyIndicator: ") {
+//                actualDuty.interestOnlyIndicator shouldBe expectedDuty.interestOnlyIndicator
+//              }
+//          }
+//        }
+//      case None         => fail("Response body is empty")
+//    }
+//  }
+//
   def checkAmountIntTotalAndCombinedDailyAccrual(
     amountIntTotal: BigInt,
     combinedDailyAccrual: BigInt,
